pyproject.toml
[project]
name = "autosim"
version = "0.0.1"
description = "A package to generate simulation data easily"
readme = "README.md"
license = { file = "LICENSE" }
authors = [
    { name = "AI for Physical Systems Team at The Alan Turing Institute", email = "ai4physics@turing.ac.uk" },
]
requires-python = ">=3.10,<3.13"

dependencies = [
    "numpy>=1.24",
    "scipy>=1.10",
    "tqdm>=4.65",
    "torch>=2.0",
    "matplotlib",
    "scikit-learn>=1.7.2",
    "einops>=0.8.2",
    "hydra-core>=1.3",
]

[project.scripts]
autosim = "autosim.cli:main"

[project.optional-dependencies]
dev = [
    "ipykernel>=7.1.0",
    "pytest>=9.0.1",
    "pytest-cov>=7.0.0",
    "ruff==0.12.11",
    "pyright==1.1.407",
    "pre-commit>=4.4.0",
]

[build-system]
requires = ["uv_build"]
build-backend = "uv_build"

[tool.coverage.run]
relative_files = true
source = [".", "/tmp"]

[tool.pytest.ini_options]
addopts = "--ignore=v0"

[tool.pyright]
venvPath = "."
venv = ".venv"

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.format]
docstring-code-format = true

[tool.ruff.lint]
select = [
    "D",   # docstring conventions
    "E",
    "F",
    "W",   # flake8
    "B",   # flake8-bugbear
    "I",   # isort
    "ARG", # flake8-unused-arguments
    "C4",  # flake8-comprehensions
    "EM",  # flake8-errmsg
    "ICN", # flake8-import-conventions
    "ISC", # flake8-implicit-str-concat
    "G",   # flake8-logging-format
    "PGH", # pygrep-hooks
    "PIE", # flake8-pie
    "PL",  # pylint
    "PT",  # flake8-pytest-style
    "RET", # flake8-return
    "RUF", # Ruff-specific
    "SIM", # flake8-simplify
    "UP",  # pyupgrade
    "YTT", # flake8-2020
    "EXE", # flake8-executable
]

ignore = [
    "PLR2004", # Magic value used in comparison
    "EM102",   # Exception must not use an f-string literal, assign to variable first
    "ISC001",  # Conflicts with formatter
    # "D417",    # Missing trailing new line in docstring
    "D100",    # Missing docstring in public module
    "D104",    # Missing docstring in public package
    "PLR0913", # too many arguments
    "RET504",
]

unfixable = [
    "F401", # Would remove unused imports
    "F841", # Would remove unused variables
]
flake8-unused-arguments.ignore-variadic-names = true # allow unused *args/**kwargs

[tool.ruff.lint.pydocstyle]
convention = "numpy"

[tool.ruff.lint.per-file-ignores]
"tests/*.py" = ["D"]
